Setup Project for 3 sum
- Created Solution class for 3 sum
- Created test class for 3 sum
