+++ src/test/java/com/lperthel/sequences/ThreeSumTest.java
@@ -1,0 +1,74 @@
+package com.lperthel.sequences;
+
+import static org.junit.Assert.fail;
+
+import org.junit.Test;
+
+public class ThreeSumTest {
+private Solution solution = new Solution();
+	@Test
+	public void testThreeSum() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testObject() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testGetClass() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testHashCode() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testEquals() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testClone() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testToString() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testNotify() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testNotifyAll() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testWait() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testWaitLong() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testWaitLongInt() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testFinalize() {
+		fail("Not yet implemented");
+	}
+
+}
